Use --current-env switch for these tests
This will use the Python interpreter that is associated
with py.test. This is likely to make these tests run on more
machines. For example, in our testing notebooks we have used
a python2 kernel. If we are not using --current-env, we would
be looking for a python2 interpreter when we run the code.