Allow checking of raised exceptions
Sometimes a notebook cell raises an exception.  Previously,
nbval would regard this as a failure for the cell.  Add the ability to
annotate a cell with "# NBVAL_RAISES_EXCEPTION".  This indicates that
we expect the cell to raise an exception.  The name of the exception
is checked against the stored value (avoiding details regarding
parsing tracebacks).