Test notebook for cells with a semicolon
This notebook contains a cell where the output is suppressed
with a semicolon. Nbval fails for this cell.

The notebook was created with the most recent version of jupyter
notebook. It was run from the top, saved and closed.

How to reproduce:

```
pytest --nbval tests/semicolon.ipynb
```

What should happen: The test passes.